+++ lib/src/widgets/scanner/overlays/standard_scanner_overlay.dart
@@ -1,7 +1,5 @@
-import 'package:basf_flutter_components/src/widgets/scanner/overlays/widgets/blackout.dart';
-import 'package:basf_flutter_components/src/widgets/scanner/overlays/widgets/scan_support_text.dart';
-import 'package:basf_flutter_components/src/widgets/scanner/overlays/widgets/toggle_direction_icon.dart';
-import 'package:basf_flutter_components/src/widgets/scanner/overlays/widgets/toggle_flash_icon_button.dart';
+import 'package:basf_flutter_components/src/widgets/scanner/overlays/widgets/scanner_blackout_square.dart';
+import 'package:basf_flutter_components/src/widgets/scanner/overlays/widgets/standard_bottom_actions.dart';
 import 'package:flutter/material.dart';
 
 /// Standard overlay for scanner
@@ -19,19 +17,12 @@ class StandardScannerOverlay extends StatelessWidget {
   Widget build(BuildContext context) {
     return Stack(
       children: [
-        const Blackout(),
+        const ScannerBlackoutSquare(),
         Positioned(
           bottom: 5,
           left: 3,
           right: 3,
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              const ToggleDirectionIconButton(),
-              ScanSupportText(scanQRorBarcode: scanQRorBarcode),
-              const ToggleFlashIconButton(),
-            ],
-          ),
+          child: StandardBottomActions(scanQRorBarcode: scanQRorBarcode),
         ),
       ],
     );