+++ lib/src/widgets/scanner/widgets/no_camera_icon.dart
@@ -3,14 +3,23 @@ import 'package:flutter/material.dart';
 /// No camera icon
 class NoCameraIcon extends StatelessWidget {
   ///
-  const NoCameraIcon({super.key});
+  const NoCameraIcon({
+    super.key,
+    this.size = 30,
+    this.color,
+  });
+
+  ///
+  final double size;
+  ///
+  final Color? color;
 
   @override
   Widget build(BuildContext context) {
     return Icon(
       Icons.no_photography_outlined,
-      size: 30,
-      color: Theme.of(context).primaryColor,
+      size: size,
+      color: color ?? Theme.of(context).primaryColor,
     );
   }
 }